--- rsocket-transport-ktor/rsocket-transport-ktor-tcp/src/jvmMain/kotlin/io/rsocket/kotlin/transport/ktor/tcp/defaultDispatcher.kt
@@ -1,21 +1,0 @@
-/*
- * Copyright 2015-2022 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.rsocket.kotlin.transport.ktor.tcp
-
-import kotlinx.coroutines.*
-
-internal actual val defaultDispatcher: CoroutineDispatcher get() = Dispatchers.IO
