--- rsocket-transport-ktor/rsocket-transport-ktor-websocket/api/rsocket-transport-ktor-websocket.api
@@ -1,7 +1,0 @@
-public final class io/rsocket/kotlin/transport/ktor/websocket/WebSocketConnection : io/rsocket/kotlin/Connection, kotlinx/coroutines/CoroutineScope {
-	public fun <init> (Lio/ktor/websocket/WebSocketSession;)V
-	public fun getCoroutineContext ()Lkotlin/coroutines/CoroutineContext;
-	public fun receive (Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
-	public fun send (Lio/ktor/utils/io/core/ByteReadPacket;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
-}
-
