+++ rsocket-transports/nodejs-tcp/src/jsMain/kotlin/io/rsocket/kotlin/transport/nodejs/tcp/internal/ext.kt
@@ -14,24 +14,26 @@
  * limitations under the License.
  */
 
-import rsocketbuild.*
+package io.rsocket.kotlin.transport.nodejs.tcp.internal
 
-plugins {
-    id("rsocketbuild.multiplatform-library")
+import org.khronos.webgl.*
+
+internal fun Socket.on(
+    onData: (data: Uint8Array) -> Unit,
+    onError: (error: Error) -> Unit,
+    onClose: (hadError: Boolean) -> Unit
+) {
+    on("data", onData)
+    on("error", onError)
+    on("close", onClose)
 }
 
-description = "rsocket-kotlin ktor integration"
-
-kotlin {
-    jvmTarget()
-    jsTarget()
-    nativeTargets()
-
-    sourceSets {
-        commonMain.dependencies {
-            api(projects.rsocketCore)
-            api(projects.rsocketTransportKtorWebsocketInternal)
-            //TODO ContentNegotiation will be here later
-        }
-    }
+internal fun createServer(
+    port: Int,
+    hostname: String,
+    onClose: () -> Unit,
+    listener: (Socket) -> Unit
+): Server = createServer(listener).apply {
+    on("close", onClose)
+    listen(port, hostname)
 }
